move RegressionSM from proxy to cache
