+++ src/iocore/cache/CMakeLists.txt
@@ -40,8 +40,7 @@ add_library(
 add_library(ts::inkcache ALIAS inkcache)
 
 if(BUILD_REGRESSION_TESTING)
-  target_sources(inkcache PRIVATE CacheTest.cc)
-  target_link_libraries(inkcache PRIVATE ts::proxy)
+  target_sources(inkcache PRIVATE CacheTest.cc RegressionSM.cc)
 endif()
 
 target_include_directories(inkcache PRIVATE ${CMAKE_SOURCE_DIR}/lib)
