+++ src/proxy/CMakeLists.txt
@@ -39,10 +39,6 @@ add_library(
 )
 add_library(ts::proxy ALIAS proxy)
 
-if(BUILD_REGRESSION_TESTING)
-  target_sources(proxy PRIVATE RegressionSM.cc)
-endif()
-
 target_link_libraries(
   proxy
   PUBLIC ts::inkcache ts::inkevent ts::tsutil ts::tscore
